feat(ar): add component for `AugmentedRealityView`
- added `AugmentedRealityView` to make the code
  more readable
- added a control to the return section of the
  `AugmentedRealityView` component to show the
  3D object directly in case the target image is not
  loaded properly
- added ViroSoundAnd3DObject component to avoid
  code repetition
- added an alert to inform users if `vrx` or `png` files
  are not loaded properly so that the application does
  not crash

SVA-633
