+++ src/components/augmentedReality/index.js
@@ -1,6 +1,7 @@
 export * from './ARModal';
 export * from './ARObjectList';
 export * from './AugmentedReality';
+export * from './AugmentedRealityView';
 export * from './HiddenModalAlert';
 export * from './IconForDownloadType';
 export * from './WhatIsARButton';
